+++ xmlui/src/components/ComponentProvider.tsx
@@ -21,7 +21,6 @@ import { localLinkComponentRenderer } from "./Link/Link";
 import { treeComponentRenderer } from "./Tree/TreeComponent";
 import { treeDisplayComponentRenderer } from "./TreeDisplay/TreeDisplay";
 import { buttonComponentRenderer } from "./Button/Button";
-import { helloWorldComponentRenderer } from "./HelloWorld/HelloWorld";
 import {
   h1ComponentRenderer,
   h2ComponentRenderer,
@@ -254,7 +253,6 @@ import { legendComponentRenderer } from "./Charts/Legend/Legend";
 import { lineChartComponentRenderer } from "./Charts/LineChart/LineChart";
 import { pieChartComponentRenderer } from "./Charts/PieChart/PieChart";
 
-import { editorComponentRenderer } from "./TableEditor/TableEditor";
 import { paginationComponentRenderer } from "./Pagination/Pagination";
 import { tooltipComponentRenderer } from "./Tooltip/Tooltip";
 
@@ -401,12 +399,6 @@ export class ComponentRegistry {
     }
     if (process.env.VITE_USED_COMPONENTS_Button !== "false") {
       this.registerCoreComponent(buttonComponentRenderer);
-    }
-    if (process.env.VITE_USED_COMPONENTS_HelloWorld !== "false") {
-      this.registerCoreComponent(helloWorldComponentRenderer);
-    }
-    if (process.env.VITE_USED_COMPONENTS_TableEditor !== "false") {
-      this.registerCoreComponent(editorComponentRenderer);
     }
     if (process.env.VITE_USED_COMPONENTS_Checkbox !== "false") {
       this.registerCoreComponent(checkboxComponentRenderer);