+++ xmlui/src/components/NumberBox/NumberBox.tsx
@@ -139,6 +139,10 @@ export const NumberBoxMd = createMetadata({
     },
   },
   themeVars: parseScssVar(styles.themeVars),
+  defaultThemeVars: {
+    [`paddingVertical-${COMP}`]: "$space-2",
+    [`paddingHorizontal-${COMP}`]: "$space-2",
+  },
 });
 
 export const numberBoxComponentRenderer = createComponentRenderer(